'&' in s modifier
